Make gc branch compile on Win32.
Just compile, there are still crashes.
And not even close to compiling successfully on Win64.

Review URL: http://codereview.chromium.org/7277038

git-svn-id: http://v8.googlecode.com/svn/branches/experimental@8446 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ gc/src/heap.cc
@@ -4564,9 +4564,9 @@ void Heap::IterateStrongRoots(ObjectVisitor* v, VisitMode mode) {
 // TODO(1236194): Since the heap size is configurable on the command line
 // and through the API, we should gracefully handle the case that the heap
 // size is not big enough to fit all the initial objects.
-bool Heap::ConfigureHeap(intptr_t max_semispace_size,
-                         intptr_t max_old_gen_size,
-                         intptr_t max_executable_size) {
+bool Heap::ConfigureHeap(int max_semispace_size,
+                         int max_old_gen_size,
+                         int max_executable_size) {
   if (HasBeenSetup()) return false;
 
   if (max_semispace_size > 0) {
