+++ gc/src/heap.h
@@ -352,9 +352,9 @@ class Heap {
  public:
   // Configure heap size before setup. Return false if the heap has been
   // setup already.
-  bool ConfigureHeap(intptr_t max_semispace_size,
-                     intptr_t max_old_gen_size,
-                     intptr_t max_executable_size);
+  bool ConfigureHeap(int max_semispace_size,
+                     int max_old_gen_size,
+                     int max_executable_size);
   bool ConfigureHeapDefault();
 
   // Initializes the global object heap. If create_heap_objects is true,
